@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.18.6 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,9 +27,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  cloud_firestore:
-  firebase_core: ^0.5.3
+  cupertino_icons: ^1.0.5
+  cloud_firestore: ^4.8.0
+  firebase_core: ^2.13.1
 
 dev_dependencies:
   flutter_test:
